- move test package - fix typo
