+++ build.gradle
@@ -5,7 +5,7 @@ plugins {
     id "com.ewerk.gradle.plugins.querydsl" version "1.0.10"
 }
 
-group = 'com.kakaopay'
+group = 'com.commerce'
 version = '0.0.1-SNAPSHOT'
 sourceCompatibility = '1.8'
