+++ src/test/java/com/commerce/interview/interview/business/pay/controller/PayControllerTest.java
@@ -1,11 +1,14 @@
-package com.kakaopay.interview.business.order.controller;
+package com.commerce.interview.interview.business.pay.controller;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
+import com.commerce.interview.business.claim.dto.ClaimDto;
+import com.commerce.interview.business.claim.entity.Claim;
+import com.commerce.interview.business.claim.service.ClaimService;
 import com.commerce.interview.business.member.entity.Member;
 import com.commerce.interview.business.member.repository.MemberRepository;
 import com.commerce.interview.business.order.dto.OrderDto;
 import com.commerce.interview.business.order.entity.Order;
 import com.commerce.interview.business.order.service.OrderService;
+import com.commerce.interview.business.pay.entity.Pay;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
@@ -24,7 +27,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @SpringBootTest
 @AutoConfigureMockMvc
-public class OrderControllerTest {
+public class PayControllerTest {
     private final Logger logger = LoggerFactory.getLogger(this.getClass());
 
     @Autowired
@@ -36,18 +39,24 @@ public class OrderControllerTest {
     @Autowired
     private OrderService orderService;
 
+    @Autowired
+    private ClaimService claimService;
+
     private Member member;
     private Order order;
+    private Pay pay;
+    private Claim claim;
+
 
     @BeforeEach
     void setUp() throws Exception {
         this.member = insertMember();
         this.order = insertOrder();
+        this.claim = insertClaim();
     }
 
     public Member insertMember() {
         if (this.member != null) return this.member;
-
         Member member = new Member("user", "1234", "chszard@gmail.com", "ROLE_USER", true);
         return memberRepository.save(member);
     }
@@ -64,10 +73,19 @@ public class OrderControllerTest {
         return orderService.createOrder(member, paymentDto);
     }
 
+    public Claim insertClaim() throws Exception {
+        ClaimDto.CancelDto cancelDto = ClaimDto.CancelDto.builder()
+                .cancelTotalAmt(1000L)
+                .cancelVatAmt(0L)
+                .orderNo(this.order.getOrderNo())
+                .build();
+        return claimService.cancelOrder(member, cancelDto);
+    }
+
     @Test
-    void listByOrderNo() throws Exception {
+    void listByMember() throws Exception {
         RequestBuilder request = MockMvcRequestBuilders
-                .get("/v1/" + String.valueOf(member.getMemberNo()) + "/order/" + String.valueOf(order.getOrderNo()))
+                .get("/v1/" + String.valueOf(member.getMemberNo()) + "/pay/list")
                 .accept(MediaType.APPLICATION_JSON)
                 .contentType(MediaType.APPLICATION_JSON);
 
@@ -79,26 +97,38 @@ public class OrderControllerTest {
     }
 
     @Test
-    void createOrder() throws Exception {
-
-        Long memberNo = member.getMemberNo();
-        OrderDto.PaymentDto paymentDto = OrderDto.PaymentDto.builder()
-                .cardNo("1234567890")
-                .cvc("030")
-                .expirationDate("0426")
-                .monthlyPayment(0)
-                .totalAmt(11000L)
-                .vatAmt(0L)
-                .build();
-
-        ObjectMapper mapper = new ObjectMapper();
-        String jsonString = mapper.writeValueAsString(paymentDto);
-
+    void listByOrderNo() throws Exception {
         RequestBuilder request = MockMvcRequestBuilders
-                .post("/v1/" + String.valueOf(memberNo) + "/order/create")
-                .content(jsonString)
-                .contentType(MediaType.APPLICATION_JSON)
-                .accept(MediaType.APPLICATION_JSON);
+                .get("/v1/" + String.valueOf(member.getMemberNo()) + "/pay/" + String.valueOf(order.getOrderNo() + "/list"))
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON);
+
+        MvcResult result = mockMvc.perform(request)
+                .andExpect(status().isOk())
+                .andDo(print())
+                .andReturn();
+        logger.debug("[RESULT]: " + result.getResponse().getContentAsString());
+    }
+
+    @Test
+    void getPay() throws Exception {
+        RequestBuilder request = MockMvcRequestBuilders
+                .get("/v1/" + String.valueOf(member.getMemberNo()) + "/pay/" + String.valueOf(20L))
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON);
+
+        MvcResult result = mockMvc.perform(request)
+                .andExpect(status().is4xxClientError())
+                .andDo(print())
+                .andReturn();
+    }
+
+    @Test
+    void getPayCancel() throws Exception {
+        RequestBuilder request = MockMvcRequestBuilders
+                .get("/v1/" + String.valueOf(member.getMemberNo()) + "/pay/cancel/" + String.valueOf(claim.getPay().getPayNo()))
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON);
 
         MvcResult result = mockMvc.perform(request)
                 .andExpect(status().isOk())
