+++ src/test/java/com/commerce/interview/interview/total/TotalIntegrateSenarioTest2.java
@@ -1,4 +1,4 @@
-package com.kakaopay.interview.total;
+package com.commerce.interview.interview.total;
 
 import com.commerce.interview.business.claim.dto.ClaimDto;
 import com.commerce.interview.business.claim.entity.Claim;
@@ -18,7 +18,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
 @AutoConfigureMockMvc
-public class TotalIntegrateSenarioTest1 {
+public class TotalIntegrateSenarioTest2 {
     static final Logger logger = LoggerFactory.getLogger(TotalIntegrateSenarioTest1.class);
 
     @Autowired
@@ -40,14 +40,10 @@ public class TotalIntegrateSenarioTest1 {
     }
 
     @Test
-    public void totalPaymentTest1() throws Exception {
+    public void totalPaymentTest2() throws Exception {
         logger.info("[CLAIM][START] ==== claim success1");
         insertClaim_Success1();
         logger.info("[CLAIM][PASS] ==== claim success1");
-
-        logger.info("[CLAIM][START] ==== claim success2");
-        insertClaim_Success2();
-        logger.info("[CLAIM][PASS] ==== claim success2");
 
         logger.info("[CLAIM][START] ==== claim fail1");
         try {
@@ -57,26 +53,10 @@ public class TotalIntegrateSenarioTest1 {
         }
         logger.info("[CLAIM][PASS] ==== claim fail1");
 
-        logger.info("[CLAIM][START] ==== claim fail2");
-        try {
-            insertClaim_fail2();
-        } catch (Exception e) {
-            logger.info("[CLAIM][LOG]: " + e.getMessage());
-        }
-        logger.info("[CLAIM][PASS] ==== claim fail2");
+        logger.info("[CLAIM][START] ==== claim success2");
+        insertClaim_Success2();
+        logger.info("[CLAIM][PASS] ==== claim success2");
 
-        logger.info("[CLAIM][START] ==== claim success3");
-        insertClaim_Success3();
-        logger.info("[CLAIM][PASS] ==== claim success3");
-
-
-        logger.info("[CLAIM][START] ==== claim fail3");
-        try {
-            insertClaim_fail3();
-        } catch (Exception e) {
-            logger.info("[CLAIM][LOG]: " + e.getMessage());
-        }
-        logger.info("[CLAIM][PASS] ==== claim fail3");
     }
 
     public Member insertMember() {
@@ -93,25 +73,16 @@ public class TotalIntegrateSenarioTest1 {
                 .cvc("030")
                 .expirationDate("0426")
                 .monthlyPayment(0)
-                .totalAmt(11000L)
-                .vatAmt(1000L)
+                .totalAmt(20000L)
+                .vatAmt(909L)
                 .build();
         return orderService.createOrder(member, paymentDto);
     }
 
     public Claim insertClaim_Success1() throws Exception {
         ClaimDto.CancelDto cancelDto = ClaimDto.CancelDto.builder()
-                .cancelTotalAmt(1100L)
-                .cancelVatAmt(100L)
-                .orderNo(this.order.getOrderNo())
-                .build();
-        return claimService.cancelOrder(member, cancelDto);
-    }
-
-    public Claim insertClaim_Success2() throws Exception {
-        ClaimDto.CancelDto cancelDto = ClaimDto.CancelDto.builder()
-                .cancelTotalAmt(3300L)
-                .cancelVatAmt(null)
+                .cancelTotalAmt(10000L)
+                .cancelVatAmt(0L)
                 .orderNo(this.order.getOrderNo())
                 .build();
         return claimService.cancelOrder(member, cancelDto);
@@ -119,38 +90,19 @@ public class TotalIntegrateSenarioTest1 {
 
     public Claim insertClaim_fail1() throws Exception {
         ClaimDto.CancelDto cancelDto = ClaimDto.CancelDto.builder()
-                .cancelTotalAmt(7000L)
-                .cancelVatAmt(null)
+                .cancelTotalAmt(10000L)
+                .cancelVatAmt(0L)
                 .orderNo(this.order.getOrderNo())
                 .build();
         return claimService.cancelOrder(member, cancelDto);
     }
 
-    public Claim insertClaim_fail2() throws Exception {
+    public Claim insertClaim_Success2() throws Exception {
         ClaimDto.CancelDto cancelDto = ClaimDto.CancelDto.builder()
-                .cancelTotalAmt(6600L)
-                .cancelVatAmt(700L)
+                .cancelTotalAmt(10000L)
+                .cancelVatAmt(909L)
                 .orderNo(this.order.getOrderNo())
                 .build();
         return claimService.cancelOrder(member, cancelDto);
     }
-
-    public Claim insertClaim_Success3() throws Exception {
-        ClaimDto.CancelDto cancelDto = ClaimDto.CancelDto.builder()
-                .cancelTotalAmt(6600L)
-                .cancelVatAmt(600L)
-                .orderNo(this.order.getOrderNo())
-                .build();
-        return claimService.cancelOrder(member, cancelDto);
-    }
-
-    public Claim insertClaim_fail3() throws Exception {
-        ClaimDto.CancelDto cancelDto = ClaimDto.CancelDto.builder()
-                .cancelTotalAmt(100L)
-                .cancelVatAmt(null)
-                .orderNo(this.order.getOrderNo())
-                .build();
-        return claimService.cancelOrder(member, cancelDto);
-    }
-
 }
